feat(ARCO-139): improve e2e test, rename functions
